+++ lib/commanded/commands/router.ex
@@ -606,8 +606,14 @@ defmodule Commanded.Commands.Router do
 
   @doc false
   def ensure_module_exists(module) do
-    unless Code.ensure_loaded?(module) do
-      raise "module `#{inspect(module)}` does not exist, perhaps you forgot to `alias` the namespace"
+    case Code.ensure_compiled(module) do
+      {:module, ^module} ->
+        :ok
+
+      {:error, _error} ->
+        raise ArgumentError,
+          message:
+            "module `#{inspect(module)}` does not exist, perhaps you forgot to `alias` the namespace"
     end
   end